Keep state/effect handlers closer to the state
Since we declare the "owner" of the state to be the ns, let's lift
the state handlers also to the ns level so reagent components don't
have this responsibility. Now we have a clear separation of the
responsibilities: state is the responsibility of the ns, whereas the
reagent components is solely reactive to the state and responsible to
the UI.

Putting handlers closer to the state is just for the sake of our own
sanity.